ceaser_cipher
The Caesar cipher is a simple substitution cipher in cryptography where each letter in the plaintext is shifted a certain number of places down or up the alphabet. It is named after Julius Caesar, who is historically recorded as using this encryption technique. The key in the Caesar cipher is the number of positions each letter is shifted, known as the "shift" or "key." It is a form of symmetric encryption, as the same key is used for both encryption and decryption. The Caesar cipher is easily breakable, as there are only 25 possible keys (excluding the trivial key of 0), making it susceptible to brute-force attacks.